Change tests to use a single gem layout
Currently tests use two different gem layouts at the same time. For
example:

* `tmp/test_rubygems_25698/default/`, where default gem executables are installed.
* `tmp/test_rubygems_25698/gemhome/`, where other gems are installed.

This is quite unintuitive, and does not match realworld rubygems setups,
where these locations are usually the same.

This commit changes the setup so that this location is shared. Two
tweaks are needed to get tests passing because there's a couple of
places in the code where `rubygems` does some stuff differently if
`Gem.default_dir` != `Gem.dir`. In particular,

* `bindir` is set to `File.join(Gem.dir, "bin")` instead of `RbConfig::CONFIG["bindir"]`.
* Checking for executable overwrites is skipped.

Since after make this change these locations are now the same, we need
to adapt to these behavioral differences.